backend/requirements.txt
Flask==2.3.2
requests
PyMuPDF
pdfkit
pytesseract
Pillow
